Refactor: 로또 관련 값을 가져오는 함수를 lottoUtils로 분리
