+++ src/js/utils/lottoUtils.js
@@ -1,0 +1,28 @@
+import { VALUE } from '../constants.js';
+
+export const getPriceByRank = (rank = VALUE.RANK.LOSE) => {
+  const price = {
+    [VALUE.RANK.FIRST]: VALUE.WINNING_PRICE.FIRST,
+    [VALUE.RANK.SECOND]: VALUE.WINNING_PRICE.SECOND,
+    [VALUE.RANK.THIRD]: VALUE.WINNING_PRICE.THIRD,
+    [VALUE.RANK.FOURTH]: VALUE.WINNING_PRICE.FOURTH,
+    [VALUE.RANK.FIFTH]: VALUE.WINNING_PRICE.FIFTH,
+    [VALUE.RANK.LOSE]: VALUE.WINNING_PRICE.LOSE,
+  }
+
+  return price[rank];
+};
+
+export const getRankByMatchCount = (matchCount = VALUE.MATCHED_COUNT.ZERO) => {
+  const rank = {
+    [VALUE.MATCHED_COUNT.SIX]: VALUE.RANK.FIRST,
+    [VALUE.MATCHED_COUNT.FIVE]: VALUE.RANK.THIRD,
+    [VALUE.MATCHED_COUNT.FOUR]: VALUE.RANK.FOURTH,
+    [VALUE.MATCHED_COUNT.THREE]: VALUE.RANK.FIFTH,
+    [VALUE.MATCHED_COUNT.TWO]: VALUE.RANK.LOSE,
+    [VALUE.MATCHED_COUNT.ONE]: VALUE.RANK.LOSE,
+    [VALUE.MATCHED_COUNT.ZERO]: VALUE.RANK.LOSE
+  }
+
+  return rank[matchCount];
+}
